create dedicated dir for node.js binding example
